Moved rendering test to sandbox app.
