+++ Syndra/src/Engine.h
@@ -9,3 +9,9 @@
 #include "Engine/ImGui/ImGuiLayer.h"
 
 #include "Engine/Core/Input.h"
+
+#include "Engine/Renderer/Renderer.h"
+
+#include "Engine/Renderer/Buffer.h"
+#include "Engine/Renderer/Shader.h"
+#include "Engine/Renderer/VertexArray.h"
