+++ Syndra/src/Engine/Core/Application.h
@@ -6,10 +6,6 @@
 #include "Engine/Core/Window.h"
 #include "Engine/Events/Event.h"
 #include "Engine/Events/ApplicationEvent.h"
-#include "Engine/Renderer/Buffer.h"
-#include "Engine/Renderer/Shader.h"
-#include "Engine/Renderer/VertexArray.h"
-#include "Platform/OpenGL/OpenGLBuffer.h"
 #include "Engine/Renderer/Renderer.h"
 
 namespace Syndra {
@@ -34,10 +30,6 @@ namespace Syndra {
 
 	private:
 		Ref<Window> m_window;
-		Ref<VertexArray> m_VertexArray;
-		Ref<VertexBuffer> m_VertexBuffer;
-		Ref<IndexBuffer> m_IndexBuffer;
-		Scope<Shader> m_Shader;
 		ImGuiLayer* m_ImGuiLayer;
 		
 		bool m_Running;
